+++ RELEASES.md
@@ -2,6 +2,9 @@
 
 # Unreleased
 
+- Improve uvfits time precision
+  - A second DATE group param is now used
+  - INTTIM is also used, if the time resolution was supplied
 - Remove progress bars
 - Remove mwalib-reading functions